Move BKE_python_components_id_loop to where it belongs
