+++ source/blender/blenkernel/intern/python_component.c
@@ -33,6 +33,7 @@
 
 #include "BKE_appdir.h"
 #include "BKE_context.h"
+#include "BKE_lib_query.h"
 #include "BKE_main.h"
 #include "BKE_python_component.h"
 #include "BKE_report.h"
@@ -850,4 +851,21 @@ void *BKE_python_component_argument_dict_new(PythonComponent *pc)
   return NULL;
 
 #endif /* WITH_PYTHON */
+}
+
+void BKE_python_components_id_loop(ListBase *complist, BKEPyComponentIDFunc func, void *userdata)
+{
+  PythonComponent *comp;
+
+  for (comp = complist->first; comp; comp = comp->next) {
+      ListBase *properties = &comp->properties;
+      PythonComponentProperty *prop;
+
+      for (prop = properties->first; prop; prop = prop->next) {
+#define PT_DEF(name, lower, upper) \
+          func(comp, (ID **)&prop->lower, userdata, IDWALK_CB_USER);
+          POINTER_TYPES
+#undef PT_DEF
+      }
+  }
 }
